Test the helper instead of the view
